- dont allow mobs with allowedtoentercombat to do that - immune message for spells wich are not applied cause of DR.

+++ trunk_4574/src/ascent-world/AIInterface.cpp
@@ -1290,7 +1290,7 @@ void AIInterface::AttackReaction(Unit* pUnit, uint32 damage_dealt, uint32 spellI
 		pUnit = m_Unit->GetMapMgr()->GetUnit(static_cast<Player *>(pUnit)->GetMisdirectionTarget());
 	}
 
-	if( m_AIState == STATE_IDLE || m_AIState == STATE_FOLLOWING )
+	if( (m_AIState == STATE_IDLE || m_AIState == STATE_FOLLOWING) && m_Unit->GetAIInterface()->GetAllowedToEnterCombat())
 	{
 		WipeTargetList();
